refactor grouping models
